backend/pyproject.toml
[project]
name = "ltchat-backend"
version = "0.1.0"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "sqlmodel>=0.0.22",
  "asyncpg>=0.29.0",
  "pydantic-settings>=2.5.0",
  "alembic>=1.13.2",
  "psycopg[binary]>=3.2.0",
  "passlib[bcrypt]>=1.7.4",
  "python-jose[cryptography]>=3.3.0",
  "bcrypt==4.0.1",
  "python-multipart",
    "pydantic[email]",
    "redis>=5.0.0",
    "aiokafka>=0.11.0",
]

[tool.uv]
dev-dependencies = [
  "pytest", "httpx", "pytest-asyncio"
]
